+++ src/DotNext.Threading/Threading/AsyncExchanger.cs
@@ -1,0 +1,236 @@
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Runtime.CompilerServices;
+using System.Threading;
+using System.Threading.Tasks;
+using static System.Threading.Timeout;
+
+namespace DotNext.Threading
+{
+    /// <summary>
+    /// A synchronization point at which two async flows can cooperate and swap elements
+    /// within pairs.
+    /// </summary>
+    /// <remarks>
+    /// This type is useful to organize pipeline between consumer and producer.
+    /// </remarks>
+    /// <typeparam name="T">The type of objects that may be exchanged.</typeparam>
+    public class AsyncExchanger<T> : Disposable, IAsyncDisposable
+    {
+        private sealed class ExchangePoint : TaskCompletionSource<T>, IDisposable
+        {
+            // cached callback to avoid extra memory allocation
+            private static readonly Action<object> CancellationCallback;
+
+            static ExchangePoint()
+            {
+                CancellationCallback = CancellationRequested;
+
+                static void CancellationRequested(object state)
+                    => Unsafe.As<ExchangePoint>(state).CancellationRequested();
+            }
+
+            private readonly T producerResult;
+            private readonly CancellationTokenRegistration registration;
+            private readonly CancellationTokenSource? source;
+
+            internal ExchangePoint(T result, TimeSpan timeout, CancellationToken token)
+                : base(TaskCreationOptions.RunContinuationsAsynchronously)
+            {
+                if (timeout > InfiniteTimeSpan)
+                {
+                    source = token.CanBeCanceled ?
+                        CancellationTokenSource.CreateLinkedTokenSource(token) :
+                        new CancellationTokenSource();
+                    source.CancelAfter(timeout);
+                    token = source.Token;
+                }
+
+                registration = token.Register(CancellationCallback, this);
+                producerResult = result;
+            }
+
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            private void CancellationRequested()
+            {
+                var token = registration.Token;
+                if (!token.IsCancellationRequested)
+                    token = new CancellationToken(true);
+
+                TrySetCanceled(token);
+            }
+
+            internal bool TryExchange(T value, [MaybeNullWhen(false)] out T result)
+            {
+                if (TrySetResult(value))
+                {
+                    result = producerResult;
+                    return true;
+                }
+
+                result = default;
+                return false;
+            }
+
+            private void Dispose(bool disposing)
+            {
+                if (disposing)
+                {
+                    registration.Dispose();
+                    source?.Dispose();
+                }
+            }
+
+            public void Dispose()
+            {
+                Dispose(true);
+                GC.SuppressFinalize(this);
+            }
+
+            ~ExchangePoint() => Dispose(false);
+        }
+
+        [SuppressMessage("Usage", "CA2213", Justification = "The field is disposed after atomic exchange")]
+        private ExchangePoint? point;
+        private bool disposeRequested;
+        private volatile ExchangeTerminatedException? termination;
+
+        /// <summary>
+        /// Waits for another flow to arrive at this exchange point,
+        /// then transfers the given object to it, receiving its object as return value.
+        /// </summary>
+        /// <param name="value">The object to exchange.</param>
+        /// <param name="timeout">The maximum time to wait.</param>
+        /// <param name="token">The token that can be used to cancel the operation.</param>
+        /// <returns>The object provided by another async flow.</returns>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="timeout"/> is negative.</exception>
+        /// <exception cref="OperationCanceledException">The operation has been canceled or timed out.</exception>
+        /// <exception cref="ObjectDisposedException">The object has been disposed.</exception>
+        /// <exception cref="ExchangeTerminatedException">The exhange has been terminated.</exception>
+        [MethodImpl(MethodImplOptions.Synchronized)]
+        public ValueTask<T> ExchangeAsync(T value, TimeSpan timeout, CancellationToken token = default)
+        {
+            ValueTask<T> result;
+
+            if (IsDisposed)
+            {
+                result = new ValueTask<T>(GetDisposedTask<T>());
+            }
+            else if (disposeRequested)
+            {
+                Dispose();
+                result = new ValueTask<T>(GetDisposedTask<T>());
+            }
+            else if (termination != null)
+            {
+                result = new ValueTask<T>(Task.FromException<T>(termination));
+            }
+            else if (point is null)
+            {
+                point = new ExchangePoint(value, timeout, token);
+                result = new ValueTask<T>(point.Task);
+            }
+            else if (point.TryExchange(value, out var arrivedValue))
+            {
+                point.Dispose();
+                point = null;
+                result = new ValueTask<T>(arrivedValue);
+            }
+            else
+            {
+                point.Dispose();
+                point = new ExchangePoint(value, timeout, token);
+                result = new ValueTask<T>(point.Task);
+            }
+
+            return result;
+        }
+
+        /// <summary>
+        /// Waits for another flow to arrive at this exchange point,
+        /// then transfers the given object to it, receiving its object as return value.
+        /// </summary>
+        /// <param name="value">The object to exchange.</param>
+        /// <param name="token">The token that can be used to cancel the operation.</param>
+        /// <returns>The object provided by another async flow.</returns>
+        /// <exception cref="OperationCanceledException">The operation has been canceled.</exception>
+        /// <exception cref="ObjectDisposedException">The object has been disposed.</exception>
+        /// <exception cref="ExchangeTerminatedException">The exhange has been terminated.</exception>
+        public ValueTask<T> ExchangeAsync(T value, CancellationToken token = default)
+            => ExchangeAsync(value, InfiniteTimeSpan, token);
+
+        /// <summary>
+        /// Informs another participant that no more data will be exchanged with it.
+        /// </summary>
+        /// <param name="exception">The optional exception indicating termination reason.</param>
+        /// <exception cref="ObjectDisposedException">The object has been disposed.</exception>
+        /// <exception cref="InvalidOperationException">The exchange is already terminated.</exception>
+        [MethodImpl(MethodImplOptions.Synchronized)]
+        public void Terminate(Exception? exception = null)
+        {
+            ThrowIfDisposed();
+            if (termination != null)
+                throw new InvalidOperationException();
+
+            ExchangeTerminatedException tmp;
+            termination = tmp = new ExchangeTerminatedException(exception);
+            if (point?.TrySetException(tmp) ?? false)
+            {
+                point.Dispose();
+                point = null;
+            }
+        }
+
+        /// <summary>
+        /// Gets a value indicating whether this exchange has been terminated.
+        /// </summary>
+        /// <exception cref="ObjectDisposedException">The object has been disposed.</exception>
+        public bool IsTerminated
+        {
+            get
+            {
+                ThrowIfDisposed();
+                return termination != null;
+            }
+        }
+
+        /// <inheritdoc />
+        protected override void Dispose(bool disposing)
+        {
+            disposeRequested = true;
+            if (disposing)
+            {
+                var point = Interlocked.Exchange(ref this.point, null);
+                if (point != null && TrySetDisposedException(point))
+                    point.Dispose();
+                termination = null;
+            }
+
+            base.Dispose(disposing);
+        }
+
+        [MethodImpl(MethodImplOptions.Synchronized)]
+        private Task DisposeAsyncImpl()
+        {
+            disposeRequested = true;
+
+            if (point is null || point.Task.IsCompleted)
+            {
+                Dispose();
+                return Task.CompletedTask;
+            }
+
+            return point.Task.ContinueWith(SuppressFaultOrCancellation, point, CancellationToken.None, TaskContinuationOptions.ExecuteSynchronously, TaskScheduler.Current);
+
+            static void SuppressFaultOrCancellation(Task<T> task, object state)
+                => (state as IDisposable)?.Dispose();
+        }
+
+        /// <summary>
+        /// Provides graceful shutdown of this instance.
+        /// </summary>
+        /// <returns>The task representing state of asynchronous graceful shutdown.</returns>
+        public ValueTask DisposeAsync()
+            => new ValueTask(IsDisposed ? Task.CompletedTask : DisposeAsyncImpl());
+    }
+}