+++ src/DotNext.Threading/Threading/QueuedSynchronizer.cs
@@ -139,7 +139,8 @@ namespace DotNext.Threading
             where TNode : WaitNode
             where TManager : struct, ILockManager<TNode>
         {
-            ThrowIfDisposed();
+            if (IsDisposed)
+                return GetDisposedTask<bool>();
             if (timeout < TimeSpan.Zero && timeout != InfiniteTimeSpan)
                 throw new ArgumentOutOfRangeException(nameof(timeout));
             if (token.IsCancellationRequested)
@@ -236,7 +237,7 @@ namespace DotNext.Threading
                 if (lockStateChecker(synchronizer))
                     return new ValueTask(synchronizer.DisposeAsync());
 
-                Dispose();
+                synchronizer.Dispose();
                 return new ValueTask();
             }
         }