+++ src/DotNext/Threading/Tasks/Continuation.cs
@@ -2,6 +2,11 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading;
 using System.Threading.Tasks;
+using static System.Threading.Timeout;
+using static InlineIL.IL;
+using static InlineIL.IL.Emit;
+using static InlineIL.MethodRef;
+using static InlineIL.TypeRef;
 
 namespace DotNext.Threading.Tasks
 {
@@ -107,5 +112,55 @@ namespace DotNext.Threading.Tasks
 
         internal static Task AttachState(this Task task, object? state, CancellationToken token = default)
             => task.ContinueWith(WhenFaultedOrCanceledAction, state, token, TaskContinuationOptions.None, TaskScheduler.Default);
+
+        private static async Task<T> WaitAsyncImpl<T>(Task<T> task, TimeSpan timeout, CancellationToken token)
+        {
+            using (var tokenSource = token.CanBeCanceled ? CancellationTokenSource.CreateLinkedTokenSource(token) : new CancellationTokenSource())
+            {
+                if (ReferenceEquals(task, await Task.WhenAny(task, Task.Delay(timeout, tokenSource.Token)).ConfigureAwait(false)))
+                {
+                    tokenSource.Cancel();   // ensure that Delay task is cancelled
+                    return await task.ConfigureAwait(false);
+                }
+            }
+
+            token.ThrowIfCancellationRequested();
+            throw new TimeoutException();
+        }
+
+        /// <summary>
+        /// Attaches timeout and, optionally, token to the task.
+        /// </summary>
+        /// <param name="task">The source task.</param>
+        /// <param name="timeout">The timeout of the asynchronous task.</param>
+        /// <param name="token">The token that can be used to cancel the constructed task.</param>
+        /// <typeparam name="T">The type of the task.</typeparam>
+        /// <returns>The task with attached timeout and token.</returns>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="timeout"/> is negative.</exception>
+        /// <exception cref="OperationCanceledException">The operation has been canceled.</exception>
+        /// <exception cref="TimeoutException">The timeout has occurred.</exception>
+        public static Task<T> ContinueWithTimeout<T>(this Task<T> task, TimeSpan timeout, CancellationToken token = default)
+        {
+            if (timeout < TimeSpan.Zero && timeout != InfiniteTimeSpan)
+                return Task.FromException<T>(new ArgumentOutOfRangeException(nameof(timeout)));
+            if (token.IsCancellationRequested)
+                return Task.FromCanceled<T>(token);
+            if (task.IsCompleted)
+                return task;
+            if (timeout == TimeSpan.Zero)
+                return Task.FromException<T>(new TimeoutException());
+            if (timeout > InfiniteTimeSpan)
+                return WaitAsyncImpl<T>(task, timeout, token);
+            if (token.CanBeCanceled)
+            {
+                Ldnull();
+                Ldftn(PropertyGet(Type<Task<T>>(), nameof(Task<T>.Result)));
+                Newobj(Constructor(Type<Func<Task<T>, T>>(), Type<object>(), Type<IntPtr>()));
+                Pop(out Func<Task<T>, T> continuation);
+                return task.ContinueWith(continuation, token, TaskContinuationOptions.ExecuteSynchronously, TaskScheduler.Current);
+            }
+
+            return task;
+        }
     }
 }